* Changes for reducing sesame deps to the minimum.
