+++ core/ivy.xml
@@ -18,17 +18,23 @@
 	<dependencies>
 	    <dependency org="com.google.guava" name="guava" rev="17.0" conf="compile,default,test->default"/>
 
-		<dependency org="com.github.jsonld-java" name="jsonld-java" rev="0.5.0" conf="compile,default,test->default"/>
+		<!--<dependency org="com.github.jsonld-java" name="jsonld-java" rev="0.5.0" conf="compile,default,test->default"/>-->
 
-		<dependency org="org.openrdf.sesame" name="sesame-runtime" rev="2.7.12" conf="compile,default,test->default">
-			<exclude org="org.slf4j"/>
-		</dependency>
+		<dependency org="org.openrdf.sesame" name="sesame-model" rev="2.7.14" conf="compile,default,test->default"/>
+		<dependency org="org.openrdf.sesame" name="sesame-query" rev="2.7.14" conf="compile,default,test->default"/>
+		<dependency org="org.openrdf.sesame" name="sesame-queryalgebra-model" rev="2.7.14" conf="compile,default,test->default"/>
+		<dependency org="org.openrdf.sesame" name="sesame-repository-api" rev="2.7.14" conf="compile,default,test->default"/>
+		<dependency org="org.openrdf.sesame" name="sesame-sail-api" rev="2.7.14" conf="compile,default,test->default"/>
+		<dependency org="org.openrdf.sesame" name="sesame-queryparser-sparql" rev="2.7.14" conf="compile,default,test->default"/>
+		<dependency org="org.openrdf.sesame" name="sesame-queryresultio-api" rev="2.7.14" conf="compile,default,test->default"/>
+		<dependency org="org.openrdf.sesame" name="sesame-rio" rev="2.7.14" conf="compile,default,test->default"/>
 
-		<dependency org="org.slf4j" name="slf4j-api" rev="1.7.5" conf="compile,default,test->default"/>
-		<dependency org="org.slf4j" name="slf4j-jdk14" rev="1.7.5" conf="default,test->default"/>
+		<dependency org="org.openrdf.sesame" name="sesame-repository-sail" rev="2.7.14" conf="test->default"/>
+		<dependency org="org.openrdf.sesame" name="sesame-sail-memory" rev="2.7.14" conf="test->default"/>
+
+		<dependency org="org.slf4j" name="slf4j-api" rev="1.7.7" conf="compile,default,test->default"/>
+		<dependency org="org.slf4j" name="slf4j-jdk14" rev="1.7.7" conf="default,test->default"/>
 
 	    <dependency org="junit" name="junit" rev="4.8.2" conf="compile,test,default->default"/>
-	    <dependency org="org.slf4j" name="slf4j-api" rev="1.7.5" conf="compile,default,test->default"/>
-	    <dependency org="commons-dbcp" name="commons-dbcp" rev="1.3" conf="compile,default,test->default"/>
 	</dependencies>
 </ivy-module>
