+++ core/main/src/com/complexible/common/openrdf/query/util/DescribeVisitor.java
@@ -18,48 +18,8 @@ package com.complexible.common.openrdf.query.util;
 import org.openrdf.query.algebra.DescribeOperator;
 import org.openrdf.query.algebra.helpers.QueryModelVisitorBase;
 import org.openrdf.query.algebra.TupleExpr;
-import org.openrdf.query.algebra.ProjectionElem;
-import org.openrdf.query.algebra.SameTerm;
-import org.openrdf.query.algebra.ValueConstant;
-import org.openrdf.query.algebra.Join;
-import org.openrdf.query.algebra.Union;
-import org.openrdf.query.algebra.LeftJoin;
-import org.openrdf.query.algebra.Filter;
-import org.openrdf.query.algebra.SingletonSet;
-import org.openrdf.query.algebra.BinaryTupleOperator;
-import org.openrdf.query.algebra.StatementPattern;
-import org.openrdf.query.algebra.Var;
-import org.openrdf.query.algebra.Projection;
-import org.openrdf.query.algebra.MultiProjection;
-import org.openrdf.query.algebra.Extension;
-import org.openrdf.query.algebra.ProjectionElemList;
-import org.openrdf.query.algebra.ExtensionElem;
-import org.openrdf.query.algebra.Or;
-import org.openrdf.query.algebra.And;
-import org.openrdf.query.algebra.BinaryValueOperator;
-import org.openrdf.query.algebra.UnaryValueOperator;
-import org.openrdf.query.algebra.FunctionCall;
-import org.openrdf.query.algebra.ValueExpr;
-import org.openrdf.query.algebra.Bound;
-import org.openrdf.query.algebra.evaluation.impl.ConstantOptimizer;
-import org.openrdf.query.algebra.evaluation.ValueExprEvaluationException;
-import org.openrdf.query.parser.ParsedQuery;
-import org.openrdf.query.parser.sparql.SPARQLParser;
-import org.openrdf.query.impl.MapBindingSet;
-import org.openrdf.query.impl.DatasetImpl;
-import org.openrdf.query.impl.EmptyBindingSet;
-import org.openrdf.query.QueryEvaluationException;
-import org.openrdf.model.Value;
-import org.openrdf.model.Literal;
-import org.openrdf.model.impl.ValueFactoryImpl;
-import org.openrdf.model.impl.BooleanLiteralImpl;
 
-import java.util.Set;
-import java.util.HashSet;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.ArrayList;
-import java.util.List;
+import org.openrdf.query.parser.ParsedQuery;
 
 /**
  * <p>TupleExprVisitor implementation that will scan a query model and see if it looks like the model for a describe
